requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
django-cors-headers==4.2.0